Protected route for profile
Note:
Profile page has not been implemented just yet
